+++ packages/simulateur/src/containers/Simulateur.tsx
@@ -48,7 +48,7 @@ function Simulateur({ state, dispatch }: Props): JSX.Element {
   const [errorMessage, setErrorMessage] = useState<string | undefined | ReactNode>(undefined)
   const { email, isAuthenticated, loading: isLoadingAuth } = useUser()
 
-  // useEffect de récupération du token et des données de la déclaration.
+  // useEffect de récupération des données de la déclaration.
   useEffect(() => {
     async function runEffect() {
       setLoading(true)
@@ -76,8 +76,9 @@ function Simulateur({ state, dispatch }: Props): JSX.Element {
               setErrorMessage(
                 <>
                   <Text>
-                    Votre email de connexion ({email}) n'est rattaché à aucun numéro Siren. Vous devez faire une demande
-                    de rattachement en remplissant le formulaire <TextLink to="/ajout-déclarant">ici</TextLink>.
+                    Votre email de connexion ({email}) n'est pas rattaché au numéro Siren de l'entreprise. Vous devez
+                    faire une demande de rattachement en remplissant le formulaire{" "}
+                    <TextLink to="/ajout-declarant">ici</TextLink>.
                   </Text>
                 </>,
               )